modif lltoa

+++ libft/ft_lltoa_base.c
@@ -6,7 +6,7 @@
 /*   By: nbascaul <nbascaul@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/23 16:56:32 by nbascaul          #+#    #+#             */
-/*   Updated: 2020/12/03 11:13:57 by nbascaul         ###   ########.fr       */
+/*   Updated: 2021/02/04 11:48:47 by nbascaul         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,8 +19,8 @@ char	*ft_lltoa_base(unsigned long long n, const char *base)
 	unsigned long long		base_len;
 	char					*ret;
 
-	div = (n < 0) ? -n : n;
-	len = (n <= 0) ? 1 : 0;
+	div = ((int)n < 0) ? -n : n;
+	len = ((int)n <= 0) ? 1 : 0;
 	base_len = (long long)ft_strlen((char *)base);
 	while (div != 0)
 	{
